requirements.txt
accelerate==0.32.0
PyYAML==6.0.1
rouge_score==0.1.2
torch==2.4.0
tqdm==4.66.4
transformers==4.44.2
trl==0.10.1
vllm==0.5.5
wandb==0.18.3